requirements.txt
click
pyyaml
requests
PyGithub
contexttimer
gitpython
retry
scipy
